fix build by loading Leaflet map only on client
